+++ src/Services/SolrCoreService.php
@@ -5,12 +5,10 @@ namespace Firesphere\SolrSearch\Services;
 use Exception;
 use Firesphere\SolrSearch\Factories\DocumentFactory;
 use Firesphere\SolrSearch\Helpers\FieldResolver;
-use Firesphere\SolrSearch\Helpers\SolrLogger;
 use Firesphere\SolrSearch\Indexes\BaseIndex;
-use Firesphere\SolrSearch\Interfaces\ConfigStore;
+use Firesphere\SolrSearch\Traits\CoreAdminTrait;
 use Firesphere\SolrSearch\Traits\CoreServiceTrait;
 use GuzzleHttp\Client as GuzzleClient;
-use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\HandlerStack;
 use LogicException;
 use ReflectionClass;
@@ -25,8 +23,6 @@ use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\SS_List;
 use Solarium\Client;
 use Solarium\Core\Client\Adapter\Guzzle;
-use Solarium\QueryType\Server\CoreAdmin\Query\Query;
-use Solarium\QueryType\Server\CoreAdmin\Result\StatusResult;
 use Solarium\QueryType\Update\Result;
 
 /**
@@ -42,6 +38,7 @@ class SolrCoreService
     use Injectable;
     use Configurable;
     use CoreServiceTrait;
+    use CoreAdminTrait;
     /**
      * Unique ID in Solr
      */
@@ -79,10 +76,6 @@ class SolrCoreService
      * @var array Valid indexes out of the base indexes
      */
     protected $validIndexes = [];
-    /**
-     * @var Query A core admin object
-     */
-    protected $admin;
 
     /**
      * SolrCoreService constructor.
@@ -133,98 +126,6 @@ class SolrCoreService
         $reflectionClass = new ReflectionClass($subindex);
 
         return $reflectionClass->isInstantiable();
-    }
-
-    /**
-     * Create a new core
-     *
-     * @param $core string - The name of the core
-     * @param ConfigStore $configStore
-     * @return bool
-     * @throws Exception
-     * @throws GuzzleException
-     */
-    public function coreCreate($core, $configStore): bool
-    {
-        $action = $this->admin->createCreate();
-
-        $action->setCore($core);
-        $action->setInstanceDir($configStore->instanceDir($core));
-        $this->admin->setAction($action);
-        try {
-            $response = $this->client->coreAdmin($this->admin);
-
-            return $response->getWasSuccessful();
-        } catch (Exception $e) {
-            $solrLogger = new SolrLogger();
-            $solrLogger->saveSolrLog('Config');
-
-            throw new Exception($e);
-        }
-    }
-
-    /**
-     * Reload the given core
-     *
-     * @param $core
-     * @return StatusResult|null
-     */
-    public function coreReload($core)
-    {
-        $reload = $this->admin->createReload();
-        $reload->setCore($core);
-
-        $this->admin->setAction($reload);
-
-        $response = $this->client->coreAdmin($this->admin);
-
-        return $response->getStatusResult();
-    }
-
-    /**
-     * Check the status of a core
-     *
-     * @deprecated backward compatibility stub
-     * @param string $core
-     * @return StatusResult|null
-     */
-    public function coreIsActive($core)
-    {
-        return $this->coreStatus($core);
-    }
-
-    /**
-     * Get the core status
-     *
-     * @param string $core
-     * @return StatusResult|null
-     */
-    public function coreStatus($core)
-    {
-        $status = $this->admin->createStatus();
-        $status->setCore($core);
-
-        $this->admin->setAction($status);
-        $response = $this->client->coreAdmin($this->admin);
-
-        return $response->getStatusResult();
-    }
-
-    /**
-     * Remove a core from Solr
-     *
-     * @param string $core core name
-     * @return StatusResult|null A result is successful
-     */
-    public function coreUnload($core)
-    {
-        $unload = $this->admin->createUnload();
-        $unload->setCore($core);
-
-        $this->admin->setAction($unload);
-        $response = $this->client->coreAdmin($this->admin);
-
-        return $response->getStatusResult();
     }
 
     /**
